Stop iff we receive 0 messages *twice* in succession
Sometimes SQS ReceiveMessage can return zero messages, even when there
are many more messages available to be consumed. This is a valid response
to ReceiveMessage, even when more messages are available, because the
contract only guarantees that MaxNumberOfMessages (default 10) _or less_
will be received each time ReceiveMessage is call.

After this change, we'll only consider the work of sqsmv to be done when
we receive zero messages twice in a row. This solution isn't perfect
(yes, there may still be more messages) but it should reduce the
likelihood of sqsmv stopping prematurely.

Closes #15